pyproject.toml
[tool.poetry]
name = "ghcp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.2"
psycopg = {extras = ["binary"], version = "^3.1.18"}
channels = "^4.0.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
